pyproject.toml
[project]
name = "rss-ai-reader"
version = "0.1.0"
description = "AI-powered RSS reader that filters and summarizes articles"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.10",
    "trafilatura>=1.6.0",
    "openai>=1.0.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiohttp>=3.9.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "aiosmtplib>=3.0.0",
    "python-dateutil>=2.8.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
rss-reader = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
